refactor: split table into folder with per-concern hooks and header
