+++ components/motion/table/types.ts
@@ -1,0 +1,64 @@
+import type { ReactNode } from "react";
+
+export type SortDirection = "asc" | "desc";
+
+export type SortState = {
+  key: string;
+  direction: SortDirection;
+};
+
+export type TableColumn<T> = {
+  /** Stable key; also the default object property read for the cell + sort value. */
+  key: string;
+  /** Header content. */
+  header: ReactNode;
+  /** Allow clicking the header to sort by this column. */
+  sortable?: boolean;
+  /** Cell text alignment. */
+  align?: "left" | "center" | "right";
+  /** Column width as a CSS length, e.g. "160px" or "20%". Omit to share remaining space equally. */
+  width?: string;
+  /** Custom cell renderer. Falls back to `row[key]`. */
+  cell?: (row: T) => ReactNode;
+  /** Value used for sorting. Falls back to `row[key]`. */
+  sortValue?: (row: T) => string | number;
+};
+
+export interface TableProps<T> {
+  data: T[];
+  columns: TableColumn<T>[];
+  /** Stable id per row, required for correct selection across sorts. Defaults to row index. */
+  getRowId?: (row: T, index: number) => string;
+  /** Render a leading checkbox column with select-all in the header. */
+  selectable?: boolean;
+  selectedRowIds?: string[];
+  defaultSelectedRowIds?: string[];
+  onSelectionChange?: (ids: string[]) => void;
+  sort?: SortState | null;
+  defaultSort?: SortState | null;
+  onSortChange?: (sort: SortState | null) => void;
+  /** Allow dragging the right edge of a header to resize that column. */
+  resizable?: boolean;
+  /** Minimum column width in px when resizing. */
+  minColumnWidth?: number;
+  onColumnResize?: (key: string, width: number) => void;
+  /** Allow dragging a header grip to reorder columns. */
+  reorderable?: boolean;
+  onColumnOrderChange?: (keys: string[]) => void;
+  /** Fixed row height in px — required for virtualization. */
+  rowHeight?: number;
+  /** Scroll viewport height in px. */
+  height?: number;
+  /** Rows rendered above/below the viewport. */
+  overscan?: number;
+  emptyState?: ReactNode;
+  className?: string;
+}
+
+/** A data row paired with its stable id. */
+export type TableRow<T> = { row: T; id: string };
+
+/** Ref map from column key to its header cell, shared across the resize/reorder hooks. */
+export type HeaderCellRefs = {
+  current: Record<string, HTMLTableCellElement | null>;
+};
