+++ components/motion/table/use-column-reorder.ts
@@ -1,0 +1,100 @@
+import {
+  type PointerEvent as ReactPointerEvent,
+  useCallback,
+  useMemo,
+  useState,
+} from "react";
+import type { HeaderCellRefs, TableColumn } from "./types";
+
+export function useColumnReorder<T>({
+  columns,
+  thRefs,
+  onColumnOrderChange,
+}: {
+  columns: TableColumn<T>[];
+  thRefs: HeaderCellRefs;
+  onColumnOrderChange?: (keys: string[]) => void;
+}) {
+  const [order, setOrder] = useState<string[]>(() =>
+    columns.map((c) => c.key),
+  );
+  const [dragKey, setDragKey] = useState<string | null>(null);
+  const [dropIndex, setDropIndex] = useState<number | null>(null);
+
+  // Apply the current order, tolerating columns added/removed at runtime.
+  const orderedColumns = useMemo(() => {
+    const byKey = new Map(columns.map((c) => [c.key, c]));
+    const seen = new Set<string>();
+    const ordered: TableColumn<T>[] = [];
+    for (const key of order) {
+      const column = byKey.get(key);
+      if (column) {
+        ordered.push(column);
+        seen.add(key);
+      }
+    }
+    for (const column of columns) {
+      if (!seen.has(column.key)) ordered.push(column);
+    }
+    return ordered;
+  }, [order, columns]);
+
+  const dropIndexFor = useCallback(
+    (clientX: number) => {
+      for (let i = 0; i < orderedColumns.length; i++) {
+        const rect =
+          thRefs.current[orderedColumns[i].key]?.getBoundingClientRect();
+        if (rect && clientX < rect.left + rect.width / 2) return i;
+      }
+      return orderedColumns.length;
+    },
+    [orderedColumns, thRefs],
+  );
+
+  const startReorder = useCallback((key: string, e: ReactPointerEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+    setDragKey(key);
+    e.currentTarget.setPointerCapture(e.pointerId);
+  }, []);
+
+  const moveReorder = useCallback(
+    (e: ReactPointerEvent) => {
+      if (!dragKey) return;
+      setDropIndex(dropIndexFor(e.clientX));
+    },
+    [dragKey, dropIndexFor],
+  );
+
+  const endReorder = useCallback(
+    (e: ReactPointerEvent) => {
+      if (e.currentTarget.hasPointerCapture(e.pointerId)) {
+        e.currentTarget.releasePointerCapture(e.pointerId);
+      }
+      if (dragKey && dropIndex !== null) {
+        const keys = orderedColumns.map((c) => c.key);
+        const from = keys.indexOf(dragKey);
+        if (from !== -1) {
+          const without = keys.filter((_, i) => i !== from);
+          let to = dropIndex;
+          if (from < to) to--;
+          without.splice(to, 0, dragKey);
+          setOrder(without);
+          onColumnOrderChange?.(without);
+        }
+      }
+      setDragKey(null);
+      setDropIndex(null);
+    },
+    [dragKey, dropIndex, orderedColumns, onColumnOrderChange],
+  );
+
+  return {
+    orderedColumns,
+    dragKey,
+    dropIndex,
+    startReorder,
+    moveReorder,
+    endReorder,
+  };
+}
