+++ components/motion/table/use-column-resize.ts
@@ -1,0 +1,84 @@
+import {
+  type PointerEvent as ReactPointerEvent,
+  useCallback,
+  useRef,
+  useState,
+} from "react";
+import type { HeaderCellRefs, TableColumn } from "./types";
+
+export function useColumnResize<T>({
+  orderedColumns,
+  thRefs,
+  minColumnWidth,
+  onColumnResize,
+}: {
+  orderedColumns: TableColumn<T>[];
+  thRefs: HeaderCellRefs;
+  minColumnWidth: number;
+  onColumnResize?: (key: string, width: number) => void;
+}) {
+  const resizeRef = useRef<{
+    key: string;
+    startX: number;
+    startWidth: number;
+  } | null>(null);
+  const [widths, setWidths] = useState<Record<string, number>>({});
+
+  const startResize = useCallback(
+    (key: string, e: ReactPointerEvent) => {
+      e.preventDefault();
+      e.stopPropagation();
+      // Freeze every column to its current pixel width so resizing one only
+      // moves the trailing spacer, never the other columns.
+      setWidths((prev) => {
+        const snapshot = { ...prev };
+        for (const column of orderedColumns) {
+          if (snapshot[column.key] == null) {
+            const measured = thRefs.current[column.key]?.getBoundingClientRect()
+              .width;
+            snapshot[column.key] = measured
+              ? Math.round(measured)
+              : minColumnWidth;
+          }
+        }
+        resizeRef.current = {
+          key,
+          startX: e.clientX,
+          startWidth: snapshot[key],
+        };
+        return snapshot;
+      });
+      e.currentTarget.setPointerCapture(e.pointerId);
+    },
+    [minColumnWidth, orderedColumns, thRefs],
+  );
+
+  const moveResize = useCallback(
+    (e: ReactPointerEvent) => {
+      const state = resizeRef.current;
+      if (!state) return;
+      const width = Math.max(
+        minColumnWidth,
+        state.startWidth + (e.clientX - state.startX),
+      );
+      setWidths((prev) => ({ ...prev, [state.key]: width }));
+    },
+    [minColumnWidth],
+  );
+
+  const endResize = useCallback(
+    (e: ReactPointerEvent) => {
+      const state = resizeRef.current;
+      resizeRef.current = null;
+      if (e.currentTarget.hasPointerCapture(e.pointerId)) {
+        e.currentTarget.releasePointerCapture(e.pointerId);
+      }
+      if (state) {
+        onColumnResize?.(state.key, widths[state.key] ?? state.startWidth);
+      }
+    },
+    [onColumnResize, widths],
+  );
+
+  return { widths, startResize, moveResize, endResize };
+}
